requirements.txt
pandas==1.4.1
numpy==1.22.3
natsort==8.1.0
pyBigWig==0.3.18
matplotlib==3.5.1
scipy==1.8.0
seaborn==0.11.2
sklearn==0.0
pysam==0.18.0
intervaltree==3.1.0
more-itertools==8.12.0
wandb==0.12.11
tqdm==4.63.0
tensorflow-gpu==2.8.0
tensorflow-probability==0.16.0
umap-learn==0.5.2
logomaker==0.8
h5py==3.6.0
